+++ packages/terra-slide-panel/tests/nightwatch/slide-panel-spec.js
@@ -1,6 +1,6 @@
 /* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
 /* eslint-disable no-unused-expressions */
-
+// eslint-disable-next-line import/no-extraneous-dependencies
 const screenshot = require('terra-toolkit').screenshot;
 
 module.exports = {
@@ -80,4 +80,3 @@ module.exports = {
     browser.expect.element('.terra-SlidePanel.terra-SlidePanel--fill').to.be.present;
   },
 };
-