+++ packages/terra-image/CHANGELOG.md
@@ -8,6 +8,12 @@ Unreleased
 * Converted SCSS variables to CSS custom properties for theming
 
 
+1.1.0 - (July 13, 2017)
+------------------
+Minor version bump
+* Updated nightwatch test scripts
+* Removed nightwatch.config file
+
 1.0.0 - (June 28, 2017)
 ------------------
 Initial stable release