+++ packages/terra-table/CHANGELOG.md
@@ -3,9 +3,15 @@ Changelog
 
 Unreleased
 ----------
+
+1.1.0 - (July 13, 2017)
+------------------
 ### Added
 * Added preventDefault to SingleSelect table row
 * Add missing props-tables
+* Updated nightwatch test scripts
+* Removed nightwatch.config file
+
 
 1.0.0 - (June 28, 2017)
 ------------------